some more 64bit instrs

+++ src/typedefs.h
@@ -2,11 +2,13 @@
 #ifndef PROTOVERSE_TYPEDEFS_H
 #define PROTOVERSE_TYPEDEFS_H
 
+#include <stdint.h>
+
 typedef unsigned char u8;
 typedef unsigned int u32;
 typedef unsigned short u16;
-typedef unsigned long long u64;
-typedef signed long long s64;
+typedef uint64_t u64;
+typedef int64_t s64;
 
 
 #endif /* PROTOVERSE_TYPEDEFS_H */
